pyproject.toml
[tool.poetry]
name = "coworking_booking"
version = "0.1.0"
description = ""
authors = ["sxkxixx <sasha.kornilov.1212@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi-jsonrpc = "^3.1.1"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
email-validator = "^2.1.1"
passlib = "^1.7.4"
peewee = "^3.17.1"
peewee-async = "^0.10.0"
aiopg = "^1.4.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.2.1"
aioredis = "^2.0.1"
aioboto3 = "^12.4.0"
python-multipart = "^0.0.9"
jinja2 = "^3.1.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
